fix(build): ensure API routes are restored after failed frontend builds
Replace chained shell commands with Node.js wrapper script that uses try-finally to guarantee API route restoration even when build fails, preserving build exit codes
